requirements.txt
pip>=23.3
wheel>=0.41.2
setuptools>=70.0.0
mysql-connector-python>=9.1.0
tabulate>=0.9.0